requirements.txt
PyMySQL==0.6.3
cffi==0.8.6
cryptography==0.6.1
ecdsa==0.11
flufl.enum==4.0.1
ipython==2.3.1
paramiko==1.15.1
pyOpenSSL==0.14
pyasn1==0.1.7
pycparser==2.10
pycrypto==2.6.1
python-etcd==0.3.2
python-gssapi==0.6.4
pytz==2014.10
six==1.8.0
urllib3==1.9.1
wsgiref==0.1.2
